+++ gamerank/rank.py
@@ -1,0 +1,10 @@
+"""Rank Nintendo Switch games based on past generation games."""
+
+
+def main():
+	"""Script entry point."""
+	pass
+
+
+if __name__ == '__main__':
+	main()